fixed loop params and added set_background function

+++ st7789v.c
@@ -209,10 +209,10 @@ void st7789v_clear(void) {
 
 }
 
-void st7789v_blink(uint16_t loops) {
+void st7789v_blink(uint8_t loops) {
     st7789_define_bounds(caset, raset);
     for (uint8_t l = 0; l < loops; l++) {
-        for (uint16_t i = 0; i < 256; i++) {
+        for (uint8_t i = 0; i < 256; i++) {
             st7789v_send_command(0x2C);
             for (uint16_t pixel = 0; pixel < ST7789V_WIDTH * ST7789V_HEIGHT; pixel++) {
                 st7789v_send_data(&i, 1);
